+ test case

+++ trunk/include/ui/IFont.h
@@ -25,9 +25,9 @@ public:
 	} CHAR_INFO;
 
 public:
-	virtual int GetLineHeight() const = 0;
+	virtual float GetLineHeight() const = 0;
 	virtual const CHAR_INFO* GetCharInfo(int nID) const = 0;
-	virtual int GetKerning(int nFirstID, int nSecondID) const = 0;
+	virtual float GetKerning(int nFirstID, int nSecondID) const = 0;
 
 protected:
 	IFont() {};
